# Reorder AddImageCheck asserts by buffer names instead of assert types. 