--- a/src/AddImageChecks.cpp
+++ b/src/AddImageChecks.cpp
@@ -25,7 +25,7 @@ public:
         Constrained,
         Proposed,
         Required,
-				TypeCheck,
+        TypeCheck,
         HostAlignment,
         HostNonNull,
     };