… test/performance/lots_of_inputs.cpp