CodeGen for halide functions with multiple inputs takes a long time. -time-pases showing 70% of the time is spent in register allocation. Even for a simple halide pipeline adding 256 inputs takes 43.sec to compile.

This PR reorders asserts into blocks acting on the same buffer. It seem to help codegen time ad running run_performance_lots_of_inputs with 512 input takes 15% less in Regallocation time.