@asaadaldien 2019-01-19
@steven-johnson Its my local clang-format! will revert those!