@asaadaldien 2019-01-25
Reordering by buffer is slightly faster:

Before:
```
Compile time with 1 inputs = 0.554068 s
Compile time with 2 inputs = 0.077662 s
Compile time with 4 inputs = 0.120946 s
Compile time with 8 inputs = 0.210784 s
Compile time with 16 inputs = 0.442599 s
Compile time with 32 inputs = 1.049370 s
Compile time with 64 inputs = 2.981606 s
Compile time with 128 inputs = 8.911081 s
```

After:
```
Compile time with 1 inputs = 0.570248 s
Compile time with 2 inputs = 0.062023 s
Compile time with 4 inputs = 0.099298 s
Compile time with 8 inputs = 0.164235 s
Compile time with 16 inputs = 0.361655 s
Compile time with 32 inputs = 0.834294 s
Compile time with 64 inputs = 2.320249 s
Compile time with 128 inputs = 6.965614 s
```