@asaadaldien 2019-01-25
Actually no compile-time improvement by doing this reordering. The real wins are coming from doing int32 cmp for type check.